+++ src/util/str_array.h
@@ -23,7 +23,7 @@ typedef struct str_array {
 str_array *new_str_array(void);
 void init_str_array(str_array *arr);
 void add_str(str_array *arr, char *str);
-char *get_str_idx(str_array *arr, int idx);
+char *get_str_idx(str_array *arr, size_t idx);
 void free_str_array(str_array *arr);
 void print_str_array(str_array *arr);
